Add 'coffee' command to create a Spring Boot project with hexagonal architecture
Introduced a new 'coffee' command in cmd/coffee.go that creates a new Spring Boot project with Gradle and Hexagonal Architecture using SPQR. The command prompts the user to provide details such as the project name and package name. Further, README.md is also updated with instructions on how to use the new 'coffee' command.